legacy/api/requirements.txt
setuptools<45.0.0
appdirs==1.4.2
beautifulsoup4==4.3.2
cairocffi==0.7.2
CairoSVG==1.0.22
cffi==1.7.0
cssselect==0.9.2
diff-match-patch==20121119
Django==1.8.3
django-admin-rangefilter==0.3.0
django-cachalot==1.0.3
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-filebrowser==3.5.7
django-filter==0.10.0
django-grappelli==2.5.7
django-import-export==0.2.7
djangorestframework==3.1.3
fuzzywuzzy==0.16.0
html5lib==0.999999999
httplib2==0.9.2
isodate==0.4.8
kitchen==1.2.5
lxml==3.6.4
mechanize==0.3.6
MySQL-python==1.2.5
mysqlclient==1.4.6
oauth2==1.9.0.post1
oauthlib==1.0.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
pycparser==2.14
pycrypto==2.6
PyJWT==1.4.0
pyparsing==2.1.10
Pyphen==0.9.4
python-magic==0.4.12
python-openid==2.2.5
python-resize-image==1.1.11
python-social-auth==0.2.14
pytz==2016.10
reportlab==3.2.0
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sqlparse==0.1.15
tablib==0.10.0
tinycss==0.3
validate-email==1.3
virtualenv==15.0.1
WeasyPrint==0.31
webencodings==0.5
xlrd==0.9.4
boto3==1.17.112
mysql-connector-python==8.0.23